Addressed all the issues mentioned in the email

+++ src/main/java/com/dwbook/phonebook/dao/ItemDAO.java
@@ -46,7 +46,7 @@ public interface ItemDAO extends Transactional<ItemDAO> {
     
     @Mapper(ItemMapper.class)
     @SqlQuery("select * from Item where merchantId=:merchantId and deletedOn is null")
-	List<Item> getItemByMerchantId(@Bind("merchantId") int merchantId);
+	List<Item> getItemByMerchantId(@Bind("merchantId") long merchantId);
 
 	//only show ones that did not have a deletedOn time stamp
     @Mapper(ItemMapper.class)
@@ -65,16 +65,16 @@ public interface ItemDAO extends Transactional<ItemDAO> {
 
     @Transaction
     @SqlUpdate("update Item set merchantId=:merchantId, title=:title, description=:description, unitPrice=:unitPrice, dailyLimit=:dailyLimit, weight=:weight, imageJson=:imageJson, feedbackJson=:feedbackJson, updatedOn = UNIX_TIMESTAMP() where id= :id and deletedOn is null")
-	void updateItem(@Bind("id") int id, @Bind("merchantId") int merchantId, @Bind("title") String title, @Bind("description") String description,
+	void updateItem(@Bind("id") long id, @Bind("merchantId") long merchantId, @Bind("title") String title, @Bind("description") String description,
 			@Bind("unitPrice") float unitPrice, @Bind("dailyLimit") int dailyLimit, @Bind("weight") int weight, @Bind("imageJson") String imageJson, @Bind("feedbackJson") String feedbackJson);
     
     @Transaction
     @SqlUpdate("update Item set deletedOn=UNIX_TIMESTAMP() where merchantId = :merchantId and deletedOn is null")
-	void deleteItemByMerchantId(@Bind("merchantId") int merchantId);
+	void deleteItemByMerchantId(@Bind("merchantId") long merchantId);
 
     @Transaction
     @SqlUpdate("update Item set deletedOn=UNIX_TIMESTAMP() where id = :id and deletedOn is null")
-    void deleteItem(@Bind("id") int id);
+    void deleteItem(@Bind("id") long id);
